FunctionApp/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pdf2image==1.17.0
python-dotenv==1.0.1
azure-storage-blob==12.20.0
azure-identity==1.16.0
Jinja2==3.1.4
